made ExDebugToolbar.Plug a plugs pipeline to split upcoming functionality
